型の変更とNULLチェックを追加した

+++ linux_6.1_mod/include/net/sclda.h
@@ -155,7 +155,7 @@ int sclda_sockaddr_to_str(struct sockaddr_storage *ss, struct sclda_iov *siov);
 struct sclda_iov *sclda_user_msghdr_to_str(
     const struct user_msghdr __user *umsg, size_t *vlen);
 struct sclda_iov *sclda_user_mmsghdr_to_str(const struct mmsghdr __user *umsg,
-                                            size_t vlen, size_t *sclda_iov_len);
+                                            unsigned int vlen, size_t *sclda_iov_len);
 int kernel_timespec_to_str(const struct __kernel_timespec __user *, char *,
                            int);
 #endif  // SCLDA_H
